requirements.txt
requests==2.31.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
